add `_post_coinit` directory subpackage - move `unknwn` into `_post_coinit` - add `_post_coinit` to `setup.cfg`
