+++ setup.cfg
@@ -22,6 +22,7 @@ python_requires = >=3.7
 
 packages =
 	comtypes
+	comtypes._post_coinit
 	comtypes.client
 	comtypes.server
 	comtypes.tools
